pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "scribebox"
version = "0.1.0"
description = "Transcribe YouTube URLs or local audio/video files to text or PDF."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Sandro Loch" }]
keywords = ["transcription", "youtube", "whisper", "fastapi", "pdf"]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "python-multipart>=0.0.9",
  "pydantic>=2.6",
  "reportlab>=4.0",
]

[project.optional-dependencies]
youtube = ["yt-dlp>=2024.7.0"]
whisper = ["openai-whisper>=20231117"]
faster-whisper = ["faster-whisper>=1.0.3"]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
  "mypy>=1.10",
  "types-reportlab>=4.0.0.20240106",
]

[project.scripts]
scribebox = "scribebox.cli:main"

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
line-length = 79

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
warn_return_any = true
no_implicit_optional = true
disallow_any_generics = true
disallow_subclassing_any = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
